requirements-core.txt
tensorflow==2.6.0
scikit-learn==0.24.2
numpy==1.19.5
tk==0.1.0
Pillow==8.3.2
